rubocop issues fixed
